+++ crates/ruma-identifiers-validation/src/user_id.rs
@@ -23,12 +23,22 @@ pub fn localpart_is_fully_conforming(localpart: &str) -> Result<bool, Error> {
         .bytes()
         .all(|b| matches!(b, b'0'..=b'9' | b'a'..=b'z' | b'-' | b'.' | b'=' | b'_' | b'/'));
 
-    // If it's not fully conforming, check if it contains characters that are also disallowed
-    // for historical user IDs. If there are, return an error.
-    // See https://spec.matrix.org/v1.4/appendices/#historical-user-ids
-    #[cfg(not(feature = "compat"))]
-    if !is_fully_conforming && localpart.bytes().any(|b| b < 0x21 || b == b':' || b > 0x7E) {
-        return Err(Error::InvalidCharacters);
+    if !is_fully_conforming {
+        // If it's not fully conforming, check if it contains characters that are also disallowed
+        // for historical user IDs. If there are, return an error.
+        // See https://spec.matrix.org/v1.4/appendices/#historical-user-ids
+        #[cfg(not(feature = "compat"))]
+        let is_invalid = localpart.bytes().any(|b| b < 0x21 || b == b':' || b > 0x7E);
+
+        // In compat mode, allow anything except `:` to match Synapse. The `:` check is only needed
+        // because this function can be called through `UserId::parse_with_servername`, otherwise
+        // it would be impossible for the input to contain a `:`.
+        #[cfg(feature = "compat")]
+        let is_invalid = localpart.as_bytes().contains(&b':');
+
+        if is_invalid {
+            return Err(Error::InvalidCharacters);
+        }
     }
 
     Ok(is_fully_conforming)